backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
opencv-python
numpy
tensorflow
ultralytics
pillow
pandas
reportlab
python-dotenv